Fields displayed
In this update, I've added all the necessary fields to languages,
diplomas and professional experience. Regarding the personal
information, I kept it to the end because it is the one which is gonna
take a lot of time. I needed to first to be sure of my results with the
other sections